fix creation of exposed action by turning executionZone to an alias for execId
SaveExposedExecutionZoneActionCommand has a property executionZone which
contains the ID of the executionZone, which is supposed to be in
AbstractExecutionZoneCommand.execId (at least there is code that expects
it).

But when creating an exposedExecutionZoneAction, only executionZone is
filled.

I selected the lazy solution of turning `executionZone` into an alias
for `execId`. That way we don't have to track down every usage of
SaveExposedExecutionZoneActionCommand.executionZone to convert it to
execId..