fix ExposedAction create/edit/execute
First attempt was 56f6f846cbde3e27b1be3c4d3acd251573ec93c1 / pull request #71, but this worked only for "create",
and broke "edit".
This was because execId was not as I thought an alias for executionZoneId, but in fact in this case (and in this case only)
an alias for exposedExecutionZoneActionId.

This is not very consistent to the rest of the system, where execId ALWAYS refers to an executionzone,
which lead to the original bug.

So we move this to actionId, and make execId to always refer to an executionzone.