fix autoupdate of log view
seems that grails 2.5 does not parse url parameters if the content-type
is 'application/json'. Content-Type does not make sense for a GET request
anyway, this was just a workaround for not being able to use "Accept" -
which just works if you enable it.

So, we should probably use "Accept" everywhere.